requirements.txt
flask
python-dotenv
openai
anthropic
google-generativeai
cryptography
authlib
requests

# Database
sqlalchemy
pymysql

# Optional: MCP Integration
# mcp
# anthropic-mcp
